API

Documentation of the functions/structs/etc. of the Jedy API.

Base

Jedy.Agent — Type

Agent(body[, body_hidden])

Basic struct for a single agent.

Fields

body – A Dict or Array data structure of attributes of the agent able to be involved in an evolutionary process.

(Optional) body_hidden – A Dict or Array data structure of attributes that will not be mutated/imitated/involved if the agent is used in an evolutionary process.

Examples

using Jedy

# An agent with a simple body of normal attributes.
simple_agent = Jedy.Agent(Dict("behaviour" => 3))

# An agent with some normal (sharable) attributes and some hidden attributes.
complex_agent = Jedy.Agent(
    Dict("behaviour" => 3, "colour" => 2),
    body_hidden=Dict("reputation") => 1
)

Jedy.run_simulation! — Function

run_simulation!(agents, fitnesses_computer, evolutionary_process!, n_epochs)

Runs a simulation.

Note

This is in-place for memory conservation, so the original agents array is mutated. Deep copy if there is a reason to preserve the original.

Input

agents::Array{Agent} – An array of Agent objects.

fitness_computer::Function – A function that takes in an Array{Agent} and returns an Array{Float64} of the same length.

evolutionary_process!::Function – An in-place function that takes in an Array{Agent} and Array{Float64} of the same length and performs an evolutionary process on the Array{Agent}.

n_epcohs::Int – The number of epochs (iterations) to run on the agents.

(Optional) logger::Function – A function that takes in an Array{Agent} and returns a Tuple{String, Array} of a csv/log/txt file name and 1D-array of values to be written each epoch. Warning: currently slightly performance-heavy.

Evolution Processes

Jedy.EvolutionProcesses.create_imitation_process — Method

create_imitation_processes(w)

Shorthand to create an imitation process function.

See Jedy.EvolutionProcesses.imitation_process! for more information.

Jedy.EvolutionProcesses.imitation_process! — Method

imitation_process!(agents, fitnesses, w)

Performs an (in-place) imitation process on some agents and their fitnesses with a given w value.

Input

agents::Array{Jedy.Agent} – An array of Agent objects.

fitnesses::Array{Float64} – An array of floats corresponding to each agent.

w::Float64 – The w factor between 0 and 1 for the imitation process.